changing the role of 'method' in routing
method used to play *exactly* the same role as service, in that both had
to match exactly from the subscription to the request. this didn't play
a real functional part, we could just as easily have a two-tuple
"service".

now it is still sent in rpc requests and publishes, but it is NOT
transmitted in ANNOUNCE messages. so the downside is that one peer
doesn't know what *methods* another peer supports, it has to just try an
rpc call and maybe get back an error (there is a new error type for when
(msg_type, service, mask, value) matches but not the method). however,
it is probably a reasonable assumption that all peers supporting a
particular service will support all the methods, that the method set for
a given service is agreed upon among all peers.

the upside is that the routing tables and ANNOUNCE/HANDSHAKE traffic
shrink dramatically, especially for the (probably most common) case of
many methods per service.

to reflect that 'method' is now *inside* the mask/value check in the
routing process, function signatures have been changed so that it comes
after mask and value in accept_* methods for creating subscriptions, and
after routing_id in RPC requests and publishes.