stability clean up
serialization of outgoing messages had been put in the sender coroutine
so that *the original objects* would sit in the queue instead of a
string copy, yielding memory savings on everything currently queued.

unfortunately there isn't a good way to handle a TypeError from the
serialization with that strategy (when it happened before it would
simply take down the sender coroutine and the hub would be effectively
dead). so this change puts string messages into the queue, allowing the
TypeError to raise to the user code that sends in an unserializable
object.

all serialization call-sites were also audited, and a few refactored so
that TypeErrors will raise in the appropriate places (all the way up to
the user in client code, and be dealt with like any other exception in
RPC handlers).