only send the request after bookkeeping is ready
this matters for the case of LocalTarget, where *in* the call to push
the request it's going to run the handler and put the response in place.
if the data structures to hold response information aren't already set
up it'll fail.

this was never an issue (and not even really a race) with remote
connections because the cooperative concurrency would guarantee that
this coroutine and only this coro would run until it explicitly switches
away, leaving no opportunity for the response to sneak in before we're
ready.