add singular RPC requests
Hubs and Clients can now provide a 'singular' kwarg to .rpc() or
.send_rpc() and the request will be directed at a single subscriber
instead of all of them (defaults to the old behavior, False).

The singular recipient selection is customizable by subclassing Hub and
implementing 'select_peer'. The default implementation uses a local
handler if there is one, otherwise chooses a random peer. Overriding
this on a Hub will also affect any connected Clients, as the target
selection takes place on the Hub. Clients nonetheless have the option to
specify singular in their .rpc() and send_rpc() methods, that boolean is
now added in as a component of the PROXY_REQUEST message type.